Move `Lease`
