+++ test/Ocelot.AcceptanceTests/LoadBalancer/RoundRobinAnalyzer.cs
@@ -1,4 +1,5 @@
 ﻿using KubeClient.Models;
+using Ocelot.LoadBalancer;
 using Ocelot.LoadBalancer.LoadBalancers;
 using Ocelot.Values;
 using System.Collections.Concurrent;
